% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixar_pnad_covid.R
\name{baixar_pnad_covid}
\alias{baixar_pnad_covid}
\title{Baixa e descompacta os microdados da PNAD COVID}
\usage{
baixar_pnad_covid(mes, caminho_pasta)
}
\arguments{
\item{mes}{Mês da PNAD COVID (número)}

\item{caminho_pasta}{A pasta (o caminho para a pasta) na qual os arquivos serão guardados (se não existir, a pasta será criada)}
}
\value{
Os arquivos de microdados (txt) na pasta designada
}
\description{
Baixa e descompacta os microdados da PNAD COVID (é preciso acesso a internet)
}
\examples{
baixar_pnad_covid(mes = 5, caminho_pasta = "./microdata");
}
