% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_populacao.R
\name{calcular_populacao}
\alias{calcular_populacao}
\title{Calcular a população}
\usage{
calcular_populacao(df_pnad)
}
\arguments{
\item{df_pnad_covid}{O dataframe com os dados da PNAD COVID (ver ler_pnad_covid)}
}
\value{
Um dataframe com 2 colunas: mês e o total da população
}
\description{
Calcular a população
}
\examples{
calcular_populacao(dados_pnad_covid)
}
\seealso{
ler_pnad_covid
}
