% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_pnad_covid.R
\name{ler_pnad_covid}
\alias{ler_pnad_covid}
\title{Carregar os microdados da PNAD COVID}
\usage{
ler_pnad_covid(mes)
}
\arguments{
\item{mes}{Mês da PNAD COVID (número)}

\item{caminho_pnadcovid}{A pasta na qual os arquivos de microdados estão guardados (em geral a mesma utilizada na função baixar_pnad_covid)}
}
\value{
Um dataframe com todos os microdados
}
\description{
Carregar (para o R) os microdados da PNAD COVID
}
\examples{
ler_pnad_covid(7, "./pasta_microdados");
}
\seealso{
baixar_pnad_covid
}
